# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для сервиса бронирований с поддержкой 2FA, уведомлений и Outbox\n\n## Возможности:\n- Регистрация и авторизация пользователей\n- Двухфакторная аутентификация (TOTP)\n- Создание и управление бронированиями\n- Система уведомлений (email, SMS, web)\n- Роли: client, staff, admin\n- Кэширование через Redis\n- Паттерн Outbox для надёжной доставки уведомлений",
        "title": "BookingGo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/UniversityProgramm/BookingGo"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/me/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список уведомлений текущего пользователя и помечает их как прочитанные",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Получить уведомления",
                "responses": {
                    "200": {
                        "description": "Список уведомлений",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/BookingGo_internal_entity.Notification"
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/me/notifications/settings": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Включает/отключает email и SMS уведомления",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Обновить настройки уведомлений",
                "parameters": [
                    {
                        "description": "Настройки уведомлений",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/BookingGo_internal_entity.NotificationSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Настройки обновлены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный формат",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "BookingGo_internal_entity.Notification": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_read": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "BookingGo_internal_entity.NotificationSettings": {
            "type": "object",
            "properties": {
                "is_email": {
                    "type": "boolean"
                },
                "is_sms": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: `Bearer \u003cyour_token\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "description": "Альтернативная аутентификация через cookie",
            "type": "apiKey",
            "name": "token",
            "in": "cookie"
        }
    }
}